api/requirements.txt
# django
# =====================
Django==3.2.13
django-filter==21.1
djangorestframework==3.13.1
drf-spectacular==0.22.1
Markdown==3.3.7

# test や lint
# =====================
tox==3.25.0
flake8==4.0.1
isort==5.10.1
black==22.3.0
mypy==0.960
